feat: Add rust releaser & cargo-toml updater
This adds supports for releasing Rust crates, by bumping their version
in `Cargo.toml`, and dependency specs in any of their local dependents.

It supports two use cases: either the repository contains a single crate
that's also a package, or it's a monorepo: a cargo workspace that has
multiple members.

In the first case, everything is delightfully simple. The singular
`Cargo.toml` has its `package.version` bumped, and we're home free.

In the second case, `release-please` should be invoked separately
for each member of the workspace, with `path`, `packageName` and
`monorepoTags` set.

For example, a Cargo workspace that looks like the following:

  - Cargo.toml
  - crates/
    - crate1/
      - Cargo.toml
    - crate2/
      - Cargo.toml

Where `crate2` depends on `crate1`, like so:

```
[package]
name = "crate2"
version = "0.1.0"

[dependencies]
crate1 = { version = "0.1.0", path = "../crate1" }
```

Whenever `release-please` is invoked with either:

  * path=crates/crate1, packageName=crate1, monorepoTags=true
  * path=crates/crate2, packageName=crate2, monorepoTags=true

Then we attempt to discover a top-level `Cargo.toml`. If it's a
workspace, we attempt to update the `Cargo.toml` for *all the crates*.

For example, if `crate1` is bumped, its `Cargo.toml` will have its
`package.version` field bumped. For `crate2`, its `Cargo.toml` will have
`dependencies.crate1.version` (or `dev-dependencies`, or
`build-dependencies`) bumped, *if and only if* that dependency spec
has a `path` component.

What does this enable?

  * When developing locally, only `dependencies.x.path` properties are
    used
  * When cargo-publishing, `dependencies.x.path` properties are ignored,
    that's why crate2's `Cargo.toml` needs to be updated, otherwise it would
    immediately fail to build.

Each crate of a monorepo can be published independently, with their own
prefixed tags. There's currently no way to trigger a release of a
specific crate without touching at least one file under its path with a
conventional commit. The behavior of the "commit-as" trailer is
untested.

While adding this functionality, I ran into a limitation of
`ReleasePR.commits`: in the base class, if a `sha` *and* a `path`
is passed, but the `sha` is for a commit that does not touch that
path, it's completely ignored, and *all commits* are returned.

That's why the `Rust` releaser adds its own `commits()` method, which
fetches two sets of commits:

  * All commits (currently, the last 100) for a given path
  * All commits up until `sha` (for any path)

It then takes the intersection of both of those sets, and that gets
passed to `generateChangelogEntry`. Without that workaround, the
generated changelog used to have duplicate entries - it would list
the last 100 conventional commits for every release, no matter what.

The cargo workspace manifest can only be at the top level. Something
like `rust/Cargo.toml` with `rust/crates/:crate/Cargo.toml` will not
work.

`Cargo.toml` edition does not perserve formatting - in Rust, the
`toml_edit` crate allows that, but in npm land, we simply don't have the
technology yet!

`Cargo.lock` is not updated. The safest way to do that would be to run
`cargo check`, but this assumes a recent rustup is installed, the
correct linker is installed (crates can opt into LLD), a C compiler is
installed (for C crates), Git credentials are properly set up (for
private Git repositories, private crate registries etc).

Updating `Cargo.lock` without going through `cargo` would result in a
lot of "reformatting noise commits". At least with the current solution,
the noise is reduced to a minimum: the next commit after a release will
have `updated-crate = new-version`, instead of a full-file,
mostly-whitespace changeset.