feat(github-release): create a draft release
Added a "draft" option which defaults to false. If true, a draft release
is created (release.draft == true and no tag is created).`tag_name` and
`target_commitish` are still associated with the release for future tag
creation when draft is changed to false. Using the web UI this is
prepopulated on the release form. Via the API it uses the previously
stored values unless the call explicitly changes them.