fix: use resolove to find imports in tests
This changes test imports to use `resolve` so that they can be run more easily from IDEs and tools
such as wallaby.js. Based on my testing, this looks like a safe change to make, and all tests pass
in VSCode, with node test, and in Wallaby.JS (running inside VSCode).