fix: use the GitHub release label to help find the correct PR
Looks for PRs with the `autorelease: tagged` label when considering release PRs. This helps prevent
creating release PRs that include every change in the repository as opposed to changes since the
last release when the main branch has been rebased or has had other issues.

The change to `findMergedReleasePR` means it is possible to manually update the tags on merged PRs
to either have them included or excluded from the list of candidates used to find the last release.

This also changes test imports to use `resolve` so that they can be run more easily from IDEs and
tools such as wallaby.js.

Based on my testing, this looks like a safe change to make, but perhaps it should be placed behind a
command line/feature flag?

Fixes #713