fix(go): support processing footers in gapics
Because we need to filter the commits found in the footer of a
commit we needed a hook to filter commits for the changelog. This
can't be done in the releaser as it only has the concept of raw
commits.Added the idea of supplying an optional commit filter to
ConventionalCommits.