refactor: stricter releaseType typing
releasers/index.js is the source of truth for available releasers. This
change strengthens that by introducing the ReleaseType as the typed
definition of what releasers are available.

Keeping the source of truth (ReleaseType) separate from actual ReleasePR
subclasses allows multiple types to point to the same subclass (e.g. the
case of "go" and "go-yoshi"). Prior to this change it would not have
been possible (via yargs cli options) to specify `--release-type go`

Since the only case for ReleasePR.releaserName is now obviated it's been
removed. If there a need in the future for a "display friendly"
representation of the releaser instance then we can add it back.