refactor: getFileContentsWithSimpleAPI with commit (#754)
The aggregate releaser will want to pass in the sha of a commit to get
the manifest and/or config at the last release commit. Augment
`getFileContentsWithSimpleAPI` to take either a branch name OR a commit
sha to pass as `ref` to `/repos/contents`. Per the docs, `ref` is:

"The name of the commit/branch/tag"

Perhaps they should have named that parameter more generically since
`ref` really refers only to branch/tag.
(https://git-scm.com/book/en/v2/Git-Internals-Git-References).