refactor(manifest): code layout, file format (#785)
The Manifest class uses a manifest.json file to track version info for
one or more packages* in a repo. The structure is an object of this
shape:

`{"<packagePath>": "<version>"}`

The manifest file must exist at least at the tip of the `defaultBranch`
even if it is empty at first. The Manifest class will attempt to find
the last merged Release PR and use the package versions set in the
manifest at that time as the current versions. Failing to find a package
version it will read the manifest at the tip of the `defaultBranch` to
see if a version has been added. "packagePath" additions are the only
change considered (changed or deleted packagePath entries are ignored).
If a version still cannot be determined then the Manifest falls back to
the individual ReleasePR instance's `defaultInitialVersion()`.

Configuration for the Manifest class is split. It needs the basic
options required to create a GitHub instance but then reads a
config.json file from the tip of the `defaultBranch` in the repository
to know what packagePaths it manages and what their specific ReleasePR
instance configuration should be. To start, only a subset of the current
ReleasePR options are supported. Also, top level ReleasePR options can
be set which will be the default values used by any configured packages.

An absurd configuration to demonstrate functionality:
```
{
  // optional package default overrides
  "release-type": "python", // default "node"
  "bump-minor-pre-major": true, // default false

  // default is conventional-changelog-conventionalcommits default
  // preset
  "changelog-sections": [
    {
      type: 'feat',
      section: 'Fancy Features Section'
    }
  ]

  // packages inheret above config but can override
  packages: {
    // path/to/package
    "py/foo": {
      // required for releasers that don't implement src lookup
      "package-name": "foopy",
      "bump-minor-pre-major": false
    },
    "node/pkg1": {
      "release-type": "node",
      "changelog-sections": [
        {
          type: 'feat',
          section: 'Node Features Section'
        },
        {
          type: 'fix',
          section: 'Node Bugs Section'
        }
      ]
    }
  }

}
```

This PR is just a skeleton to consider code layout and the
manifest/config design. The implementation to follow is:

- `pullRequest` - Get all the commits since last release and parcel
them out per package. Pass them along with current version info to a
releaser instance per package and get back the updates needed for a PR.
Update the manifest with new versions and open a PR.

- `githubRelease`: pass the single commit that is the last merged
Release PR into CommitSplit to determine which packages were updated.
Create a Release/Tag for each of these

- connect to factory/cli and profit

* In the release-please domain a package is a collection of
configuration options ("release-type", "package-name", "path") - the
unique/natural key is "path" which identifies a directory under the
repository root where relevant package metadata files live. All the
source code for this package must also have this path as an ancestor.