feat: custom version increment strategies
Allow for version incrementing strategies other than conventional
commits. The first example is "always-patch" which completely ignores
conventional commits and hard-codes a "patch" level version bump. This
strategy is useful in cases where the "major" and "minor" numbers are
set by conditions outside the package codebase but the "patch" number
is still available for representing changes as a new version.