fix: debug logging for node-workspaces plugin
When running as a GH action, there are times when the plugin does not
set the version correctly on a package. Specifically, if a package is
getting a version bump only because at least one dependency is version
bumping then we see that it only gets a "patch" bump even if there is a
specific releaseAs config key set for that package.

When I try to reproduce this running the cli locally on my mac laptop
the problem disappears.