feat(manifest): support hard-coded last-release-sha
allow specifying an arbitrary sha to represent the last release. This is
helpfull to use when trying to revert the last release PR. While you can
revert the changes from a bad release PR and delete associated tags/releases
GitHub does not allow you to delete the actual PR so release-please will
still find it and consider it the last release.