feat: github-release can customize the label applied to the pull request (#1060)
The `GitHubReleaseOptions` interface already includes a `releaseLabel` field that controls the label applied to the pull request when `github-release` is run. (When this field is undefined, the logic in `github-release.ts` falls back to `autorelease: tagged`.) If I understand Yargs correctly, this change will allow this field to be set via Yargs.