fix: Don't update dev-dependencies lacking a version key
This is a non-breaking change. If you have a version key for a
dev-dependency that has a path key, release-please will bump it
as before. If you don't have a version key (which is recommended,
and required for cargo-workspaces publish to work, see this discussion:
https://github.com/pksunkara/cargo-workspaces/pull/53/files#r735178500 )
then release-please will no longer fail when trying to process your
workspace.

Closes https://github.com/googleapis/release-please/issues/1094