fix: address false-positive matches for autorelease branch naming (#1311)
Autorelease branch naming can mistakenly match for any branch
beginning with a v.  A PR branch like
"release-please--branches--v3.3.x" will be treated as an autorelease
branch for a component called "please--branches-".

If the branch name is a valid semver, the wrong class will be chosen
to parse the branch name, and later steps will fail.  If it's not a
valid semver, then branch parsing will completely fail.

This excludes these false-positive matches by excluding branches which
start with "release-please--branches".

Closes #1310