fix: Restore v12 changelog formatting for ruby-yoshi (#1361)
Currently Ruby repos are pinned to release-please 12.x for several reasons, among them being that the release-please 13 rewrite reverted Ruby's preferred changelog formatting. This PR updates the `ruby-yoshi` strategy to reimplement this formatting as a postprocessing pass to buildReleaseNotes. Specifically:

* It standardizes on h3 for all version headers, regardless of whether it's a major/minor or patch version.
* It removes the diff link on version headers, as we prefer not to have links in the changelog.
* If a commit message includes a pull request number, conventional commits turns it into a link. We remove that link.
* We collapse vertical whitespace of two or more blank lines.

This PR also removes the custom pull request body logic in `yoshi-ruby`. We were not using that information.

The standard `ruby` strategy is unaltered, so that third parties using it are not affected.