feat: Support sequential-calls manifest field that disables concurrency when creating multiple pull requests or releases (#1401)
When running release-please on the google-cloud-ruby manifest, we're running afoul of secondary rate limits. Currently release-please runs pull request creation concurrently, and in this case, it was trying to open 33 pull requests at once which seemed to be too much. So we're introducing an option to run them sequentially using blocking calls.

This was added to the manifest config rather than the CLI options, because it seems to be a scaling-related knob that should be set per repository, rather than something that depends on how release-please is invoked.

I honestly do not know how to write a test for this. I tested manually both the sequential and concurrent cases to make sure the functionality was still correct, but I'm not sure how to test whether something happened concurrently or not.