fix: fix file tree listing for extremely large repositories (#1505)
Apparently, setting `recursive` to `true` or `false` make the get tree API attempt to return all files recursively.

This leads to potentially missing files if the repository is extremely large.

https://docs.github.com/en/rest/git/trees#get-a-tree

```
Setting this parameter to any value returns the objects or subtrees referenced by the tree specified in :tree_sha. For example, setting recursive to any of the following will enable returning objects or subtrees: 0, 1, "true", and "false". Omit this parameter to prevent recursively returning objects or subtrees.
```