fix(dart): correct bad test snapshots
Tests were previously passing because the snapshot was bad - snapshot was 0.5.0, test should have set 0.6.0, but because the dart strategy was broken, it was returning 0.5.0 and incorrectly passing. The snapshots have now been updated to correctly expect 0.6.0, and the strategy has been fixed to parse a multiline file properly and search/replace the string. A new test was also added to preserve non-standard build codes (the string following + in version), just in case someone uses it that way.

Finally, care was taken to determine if we could/should use the same updater strategy code as helm (also yaml), but js-yaml strips out comments which may have unexpected results for users. It's understood that the current regex-based strategy is brittle but it does work fine for the purpose it's meant for.