feat: added expo strategy and updater (#1646)
* feat: added expo strategy and updater

Added `strategies/expo` and `updaters/expo/app-json` for versioning
of Expo based React Native projects. This updater performs the same
functions as the `node` updater but also updates the `app.json` file
found in Expo projects.

* chore: added expo to the strategy list in the docs

* chore: updated copyright year on all expo strategy/updater files

* chore: added expo to src/factory

Added `expo` as a type available on the global factory. Also updated
the Expo SDK version number passed between the Expo strategy and updater
to use a `Version` object instead of a string.

* chore: fixed expo strategy expo sdk version test

* chore: fixed broken CLI snapshot tests

* fix: treat expo ios and android build numbers as optional

If a project has not specified the `expo.ios.buildNumber` or
`expo.android.versionCode`, don't attempt to set those.

Co-authored-by: Jeff Ching <chingor@google.com>