feat: add support for crystal lang projects
This change adds support for projects written in
[Crystal](https://crystal-lang.org/). In a Crystal project, version
information is stored in the `version` field of a file called
`shard.yml` in the root of your project.
Functionally, this new crystal strategy is the same as the pre-existing
Helm strategy, so I've used that as the basis for my implementation.

Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>