Update manifest-releaser.md
I use `release-please` in one of my projects and was at first not able to use the automatic bootstrap functionality. I had to enable the "Allow GitHub Actions to create and approve pull requests" to make it work. I thought it may be a good idea to add this hint to the docs :)