fix: draft-pull-requests config (#2241)
I discovered this issue while using `release-please-action` which builds
a config via `Manifest.fromManifest`. In this function the
`draftPullRequest` setting from the config from the config defaults was
not considered when merging with the package specific configs. As a
result, defining this at the top level of the config had no impact.

Integration with the action was tested via a fork of the
`release-please-action` which was run on a test repo[3] the results can
be seen at[4]

Fixes: #1791

[1] https://github.com/google-github-actions/release-please-action/blob/a37ac6e4f6449ce8b3f7607e4d97d0146028dc0b/src/index.ts#L108
[2] https://github.com/matthewhughes934/release-please-action/commit/07de5f392038d75f976ab0afff1ea124db275c3e
[3] https://github.com/matthewhughes934/release-please-test/commit/c55af6660b7e67974ed4d9924e4a378c3ab01a7c
[4] https://github.com/matthewhughes934/release-please-test/pull/10

Co-authored-by: Jeff Ching <chingor@google.com>