feat: add "always-update" config option
Fixes #1459

This patch adds a new option "always-update" that forces existing pull
requests to be updated even when there are no changes to the
release notes. This is useful, for example, when the repository
enforces pull requests to be up-to-date with the main branch before
they can be merged. Without this option set to `true`, that is, with the
default behavior of release-please, if the last commit made to the
main branch does not appear in the release notes, the existing pull
request branch would not be rebased.