feat: support release candidates in java
Active only when we have a prerelease suffix (-rc).

By default, after releasing a library versioned as X.Y.Z-rc(n),
a snapshot PR will set it to X.Y.Z-rc(n+1)-SNAPSHOT.

When determining the next release, the version may be
reset from `rc(m)` to `rc1` if:
 - a minor bump is determined for next release
 - a major bump is determined for next release

Otherwise, the `-SNAPSHOT` string will simply be removed.

NOTE: This is not compatible with ServicePack versioning.